Refactor Painter uses GraphicalContext instead of Game.
GraphicalContext object is now passed to Painter instead of
Game. Game does not have to explicity give GraphicalContext to Painter
everytime it makes a call to Painter. As a result, Painter should now
be initalized and given to the constructor of Game.
Additionally wrote some Javadoc for Painter, Snake and Game.