Add text element to show score to scene
Added a text element that is used to show the score on the gamescreen.
This text element is added to the scene during start of the game. The
Game class keeps track of the score. Everytime a piece of fruit is eaten
the score increases by 1 and the text is updated.